update lab4 changed

+++ lab4/Bai_tap_5/Dathuc.h
@@ -21,7 +21,7 @@ public:
     dathuc operator*(dathuc D2) const;
 
     // tinh toan da thuc khi biet duoc so nguyen s
-    double total(int s) const;
+    double total(double s) const;
     
     // cac ham so sanh, bang hon
     bool operator==(dathuc D2) const;
